requirements.txt
# Core LangChain
langchain==1.2.13
langchain-community==0.4.1
langchain-core==1.2.23
langchain-text-splitters==1.1.1

# LLM Provider
langchain-ollama==1.0.1
ollama==0.6.1

# Embeddings
sentence-transformers==5.3.0

# Database / PostgreSQL
langchain-postgres==0.0.17
pgvector==0.3.6
psycopg-binary==3.3.3
SQLAlchemy==2.0.48

# PDF
pypdf==6.9.2

# Utils
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2